Fixes Loop on City selection
Fixes a loop where _selectedProfession is set to a profession and a StepBack() is requested. SetStep(..) to Trade would push to ChooseCity as _selectedProfession was valid. New functionality forwards the gumps when calling SetProfession (City if profession, Trade is Advanced profession). When stepping back from ChooseCity, the client is stepped back appropriately to Profession or Trade, depending on _selectedProfession.

Added a default case to SetStep incase we somehow end up on an odd step.