✨ Add -hue command for quick inspection of in-game items
This command is available in the UO:Renaissance client, but was not
available in ClassicUO. It is a pared down version of the information
available in the "-info" command gump, intended for rapid inspection
of the most common "hidden" properties: item ID and hue.

Personally, I use this for organizing NPC cloth I take from brigands,
escorts, evil mages, and the vendors in Buc's Den.