Add better check for non-lootable items
Skip non-lootable items (the check below is for UO:Outlands only).
These items appear only on humanoid NPC corpses so they don't look naked.
They are always a piece of equipment and server always forces their position as 0,0,0 which is almost the same as Position.INVALID (you can't place an item in 0,0,0).
So the check works the next way: show an item if the shard tipe is not Outlands, show the item if it isn't equipment, show the item if position in container is not 0,0,0.