Allows Prof.txt to support skills with spaces in it
Prof.txt is used to define the templates during character creation. Skills like "Animal Taming" are not parsed correctly and the character is created without skills.

This fix adds support so if the template has "Animal_Taming" it will correctly load based on the name in SkillsLoader (in this case, "Animal Taming")