Add ability to specify profiles directory
Allow for a custom profile directory to be specified via a command
line parameter "profilesdirectory". By default this value is "" and the
 program behaves as normal and profile data will be accessed from:

 GameExecutableDirectory/<account>/<server>/<character name>

If non-empty, profile data will be accessed from custom directory:

 ProfileDirectory/<account>/<server>/<character name>

Additionaly, some refactoring of the code has been done so that the profiles
directory is only generated once and then shared between modules.