Extensible seat definitions
The client previously used a hard coded list of graphic ids to determine
which graphics make the character display the "sitting" graphic. Now,
that data is pulled from Data/Client/chairs.csv. This file is in the
following format:

Each row defines one chair:

graphic, directionNorth, directionWest, directionSouth, directionEast, Y
Offset, MirrorOffsetY

directions are defined as 0-7 for valid directions and -1 if that
direction is not allowed for this chair. These directions define how
characters are turned when sitting relative to the direction they are
actually facing. For instance, if a chair only supports sitting facing
north graphically, it may still allow characters who are facing
northwest or northeast to display as sitting (facing directly north)
without requiring a turn request be sent to the server.