Fix PUnicodePromptResponse
This version now produces exactly the same output in every case as the
original client. The language is a null terminated ASCII string
containing the language code (which is usually 3 characters). The
description is a little endian unicode string with no null terminator.

Before, it wasn't null terminating the language and starting the
description immediately. But it was sending big endian encoded unicode
and then tacking on an extra null byte. This works if all of the
characters being sent happen to only require 1 byte of unicode, but as
soon as you send something with data in the upper byte of a character it
falls apart.