"Ping" (send 0x73) in character select, create regardless of "no_server_ping"
The -no_server_ping command line is meant to deactivate the feature that
sends real ICMP pings to the server from the server selection screen.

But it also had the effect of deactivating the application layer 0x73
packets (also called a "Ping" in the code) sent from the character
creation screens.  That was actually not desirable.  It only led to the
client hanging if a user took too long to finish creating their character.

This change disregards -no_server_ping in the character creation screen,
and it also sends the application layer 0x73 "ping" from the character
selection screen, so that the user can also idle there safely.