Delay loading of bodyconv.def until expansion flags are known
By delaying this, we can eliminate BodyConvGroups entirely.

Additionally, instead of storing the new graphic in IndexAnimation.Graphic, just
set up the top level entry in the Index to point at the right stuff.
This avoids having to chase through Graphics in a loop to do body
conversions.

The only ugly part is that some of the logic to do action mapping does
depend on knowing the "original" animation type and the "new" animation
type, so we still have to store the original type.