Fix "Use old animations" flag
On an old server I ran, I had a player command to enable old animations. Here is a snippet:

```cs
					bool turnon = e.Arguments[0].ToLower() == "on";
					acct.SFO = turnon ? (FeatureFlags)0x03 : (FeatureFlags)0x0;
					m.SendMessage( "Old style client packets has been turned {0}.", turnon ? "ON" : "OFF" );
					m.Send( new MobileIncoming( m, m ) );
					m.Send( SupportedFeatures.Instantiate( m.NetState ) );
					m.Send( new MobileUpdate( m ) );
					m.Send( new MobileAttributes( m ) );
```

Note: RunUO removes the T2A, Ren, and UOTD flags for Pre-AOS era causing the same effect. The UOTD/LBR flags do not need to be enabled for this as far as I remember.