Fix Map Zones on Linux (#1608)
Here is a **naive** attempt to fix bug #1608, map zones on Linux.

From my investigation, it appears a library called `System.ServiceModel.Internals` is _not_ 
ncluded with the build. In Windows, perhaps the global system library is used, but that
appears not to be the case in Linux. I'm not sure where the build decides to bundle DDLs,
but I _think_ I've found the spot, in `tools/ManifestCreator/Program.cs`.

I unfortunately have no way of testing this bug. Thanks!