Fix Server Ping in Server List
The System.Net.NetworkInformation.Ping class requires the IP address in little-endian format. Previously, it was passed in big-endian format (IP address in reverse order. 1.2.3.4 became 4.3.2.1). As a result, the servers were never pinged correctly. To save the IP address in the correct order in _ipAddress, the byte order must be reversed so that the least significant byte (LSB) comes first.