Fix race condition when handling plugins packets
There's a race condition in the access of the packets CircularBuffers,
between the render thread, which reads the packets from the Server->Client buffer,
the Plugin->Client buffer and the Plugins thread which can enqueue new
packets to be read by the Client.

Add a lock of the Handler in the render thread.