One of the reflection shims seem to have been forgotten in the bootstrapper
ClassicAssist was broken for quite some time but degraded gracefully by sending the pathfind packet to the client instead. That packet does not allow setting the distance however and breaks when trying to pathfind to a location that is impassable (for example pathfinding to a blocking item, like a loom or a spinning wheel). This makes pathfinding from ClassicAssist in those cases impossible